Refactor users #97
