+++ src/entity/group.entity.ts
@@ -10,7 +10,7 @@ import {
   PrimaryGeneratedColumn,
 } from 'typeorm'
 
-import { User } from './user.entity'
+import { User } from '../users/user.entity'
 
 @Entity()
 export class Group extends BaseEntity {
