+++ src/types/index.d.ts
@@ -1,5 +1,6 @@
-import { Ticket } from '../tickets/ticket.entity'
 import { Group } from '../entity/group.entity'
+import { Ticket } from '../tickets/ticket.entity'
+import { User as LocalUser } from '../users/user.entity'
 
 declare global {
   namespace Express {
@@ -8,6 +9,7 @@ declare global {
       ticket: Ticket
       groups: Group[]
       group: Group
+      userToShow: LocalUser
     }
   }
 }
