+++ views/partials/header.pug
@@ -21,5 +21,5 @@ nav.navbar(role='navigation' aria-label='main navigation')
           .navbar-link
             strong= user.name
           .navbar-dropdown
-            a.navbar-item(href="/users/me") Profil
+            a.navbar-item(href=`/users/${user.id}`) Profil
             a.navbar-item(href="/logout") Kijelentkezés
