+++ views/user/show.pug
@@ -1,16 +1,16 @@
 extends ../layouts/main
 
 block content
-  h1.title= otherUser.name
-  if user.id !== otherUser.id && user.admin
-    form(action=`/users/${otherUser.id}/admin`, method="post")
-      if otherUser.admin
+  h1.title= userToShow.name
+  if user.id !== userToShow.id && user.admin
+    form(action=`/users/${userToShow.id}/admin`, method="post")
+      if userToShow.admin
         button.button.is-danger(type="submit") Admin jog megvonása
       else
         button.button.is-link(type="submit") Admin jog adása
   h2.subtitle Csoportok
   ul
-    each group in otherUser.groups
+    each group in userToShow.groups
       li
         a(href=`/groups/${group.id}`)
           p= group.name
